requirements.txt
paho-mqtt
Pillow==7.1.2
